Move value classes to a separate package
