+++ src/main/kotlin/sigma/ArgumentTable.kt
@@ -1,5 +1,9 @@
 package sigma
 
+import sigma.values.Symbol
+import sigma.values.tables.Table
+import sigma.values.Value
+
 class ArgumentTable(
     private val name: Symbol,
     private val value: Value,
