+++ src/main/kotlin/sigma/Thunk.kt
@@ -1,5 +1,7 @@
 package sigma
 
+import sigma.values.Value
+
 interface Thunk {
     fun obtain(): Value
 }
