+++ src/main/kotlin/sigma/expressions/DictConstructor.kt
@@ -1,7 +1,7 @@
 package sigma.expressions
 
-import sigma.DictAssociativeTable
-import sigma.Table
+import sigma.values.tables.DictAssociativeTable
+import sigma.values.tables.Table
 import sigma.parser.antlr.SigmaParser
 import sigma.parser.antlr.SigmaParser.DictArrayAltContext
 import sigma.parser.antlr.SigmaParser.DictContext
