+++ src/main/kotlin/sigma/expressions/IntLiteral.kt
@@ -1,9 +1,9 @@
 package sigma.expressions
 
-import sigma.IntValue
-import sigma.Symbol
-import sigma.Table
-import sigma.Value
+import sigma.values.IntValue
+import sigma.values.Symbol
+import sigma.values.tables.Table
+import sigma.values.Value
 import sigma.parser.antlr.SigmaParser
 
 data class IntLiteral(
