+++ src/main/kotlin/sigma/expressions/LetExpression.kt
@@ -1,8 +1,8 @@
 package sigma.expressions
 
 import sigma.LoopedScope
-import sigma.Table
-import sigma.Value
+import sigma.values.tables.Table
+import sigma.values.Value
 import sigma.parser.antlr.SigmaParser.LetExpressionContext
 
 data class LetExpression(
