+++ src/main/kotlin/sigma/values/BoolValue.kt
@@ -1,4 +1,4 @@
-package sigma
+package sigma.values
 
 data class BoolValue(
     val value: Boolean,
