+++ src/main/kotlin/sigma/values/ComputableFunctionValue.kt
@@ -1,4 +1,4 @@
-package sigma
+package sigma.values
 
 // Thought: Is this needed anymore?
 abstract class ComputableFunctionValue : FunctionValue()
