+++ src/main/kotlin/sigma/values/Symbol.kt
@@ -1,4 +1,4 @@
-package sigma
+package sigma.values
 
 data class Symbol(
     val name: String,
