+++ src/main/kotlin/sigma/values/tables/DictAssociativeTable.kt
@@ -1,6 +1,7 @@
-package sigma
+package sigma.values.tables
 
 import sigma.expressions.Expression
+import sigma.values.Value
 
 class DictAssociativeTable(
     override val environment: Table,
