+++ src/main/kotlin/sigma/values/tables/EmptyTable.kt
@@ -1,4 +1,6 @@
-package sigma
+package sigma.values.tables
+
+import sigma.values.Value
 
 object EmptyTable : Table() {
     override fun read(argument: Value): Value? = null
